pointercal: Always create a package but empty on devices without a default
git-svn-id: https://svn.o-hand.com/repos/poky/trunk@1751 311d38ba-8fff-0310-9ca6-ca027cbcb966

+++ meta/packages/pointercal/pointercal_0.0.bb
@@ -1,14 +1,17 @@
 DESCRIPTION = "Touchscreen calibration data"
 SECTION = "base"
-
-COMPATIBLE_MACHINE = "(spitz|c7x0|nokia800|akita)"
+PR = "r1"
 
 SRC_URI = "file://pointercal"
 S = "${WORKDIR}"
 
+ALLOW_EMPTY_${PN} = "1"
 PACKAGE_ARCH = "${MACHINE_ARCH}"
 
 do_install() {
-        install -d ${D}${sysconfdir}/
-        install -m 0644 ${S}/pointercal ${D}${sysconfdir}/
+	# Only install file if it has a contents
+	if [ -s ${S}/pointercal ]; then
+	        install -d ${D}${sysconfdir}/
+	        install -m 0644 ${S}/pointercal ${D}${sysconfdir}/
+	fi
 }
